requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
groq>=0.4.1
python-dotenv>=1.0.0
python-anvil>=1.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.9